fix int_ and errors in LC_analytical

+++ development/RPLB_acc_NoSTC_arbitrary.py
@@ -39,7 +39,7 @@ def RPLB_acc_NoSTC_arbitrary(lambda_0, tau_0, a, P, PM, phi_2, t_0, z_0, beta_0)
     # do 5th order Adams-Bashforth finite difference method
     for k in range(0, len(time)-1):
         
-        alpha = np.linspace(0, 1.0, np.int(501+10*np.round(np.abs(z[k])/a)))
+        alpha = np.linspace(0, 1.0, np.int_(501+10*np.round(np.abs(z[k])/a)))
         d_alpha = alpha[1]-alpha[0]
         scaling = np.sqrt(2*k_0*a)*np.tan(alpha/2)
         illum = scaling*np.exp(-scaling**2)
@@ -81,7 +81,7 @@ def RPLB_acc_NoSTC_arbitrary(lambda_0, tau_0, a, P, PM, phi_2, t_0, z_0, beta_0)
 
         KE[k+1] = ((1/np.sqrt(1-beta[k+1]**2))-1)*m_e*c**2/q_e
         
-        if (time[k] > 300*tau_0 and np.mean(np.abs(np.diff(KE[k-np.int(10*n):k+1]))/(KE[k+1]*dt)) < 1e7):
+        if (time[k] > 300*tau_0 and np.mean(np.abs(np.diff(KE[k-np.int_(10*n):k+1]))/(KE[k+1]*dt)) < 1e7):
             k_stop = k+1
             break
